tricks to save space

+++ keyboards/keebio/kbo5000/keymaps/henryhu/anim.h
@@ -3,6 +3,6 @@
 #include "options.h"
 
 #ifdef ENABLE_NEKO
-void pick_frame(uint32_t idle_time);
+void pick_frame(uint16_t idle_time);
 void render_animation(uint8_t col, uint8_t row);
 #endif
