+++ keyboards/keebio/kbo5000/keymaps/henryhu/cmdmode.h
@@ -10,6 +10,6 @@
 extern bool command_mode;
 extern char cmdRet[64];
 
-bool command_mode_key(uint16_t keycode, keyrecord_t *record);
+bool command_mode_key(uint8_t keycode, keyrecord_t *record);
 void cmd_exit(char* cmd, char* buf, int size);
 #endif
